Add files via upload
I created an excel file to go with the script to get someone started on analysis. All data was freely available and I have collected over a couple different days of running the application to get information in seconds. I provided some free analysis (simply % change over a measuring period and maximum % change over the entire range of data collection). The file is formatted to work with the new AppleScript edit.