Update (。··)_且 CRYPTO(。··)_且 _APP
Updated AppleScript that has several more random currencies that remain fairly constant over the interval. The code is not perfect and sometimes a currency won't be fetched but so far bitcoin has always responded to the function. I wanted to collect a larger set but the original idea cause a table overflow error from calling too many variables. The proposed idea could maybe be done in python but I'm going to stick with AppleScript for now.